+++ src/main/java/com/example/webshop/Shopping/AddToCartFacade.java
@@ -4,7 +4,14 @@ import com.example.webshop.*;
 import com.example.webshop.inventory.InventoryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import java.math.BigDecimal;
 
+
+/**
+ * Facade service that handles operations related to adding products to the shopping cart.
+ * This service interacts with the product, shopping cart, and inventory services to manage
+ * product availability and cart updates.
+ */
 @Service
 public class AddToCartFacade {
 
@@ -12,6 +19,13 @@ public class AddToCartFacade {
     private final ShoppingCartService shoppingCartService;
     private final ProductService productService;
 
+    /**
+     * Constructor that injects dependencies for product, shopping cart, and inventory services.
+     *
+     * @param inventoryService The service responsible for managing inventory (stock levels).
+     * @param shoppingCartService The service responsible for handling shopping cart operations.
+     * @param productService The service responsible for managing product information.
+     */
     @Autowired
     public AddToCartFacade(InventoryService inventoryService, ShoppingCartService shoppingCartService, ProductService productService) {
         this.inventoryService = inventoryService;
@@ -19,27 +33,45 @@ public class AddToCartFacade {
         this.productService = productService;
     }
 
-    // Method to add product to the shopping cart with stock validation
+    /**
+     * Adds a product to the shopping cart if sufficient stock is available. If the product is
+     * out of stock, an exception is thrown.
+     *
+     * @param productId The ID of the product to be added to the cart.
+     * @return The updated shopping cart after the product is added.
+     * @throws RuntimeException if the product is not found or there is insufficient stock.
+     */
     public ShoppingCart addToCart(Long productId) {
         ProductModel product = productService.getProductById(productId);
 
+        // Check if the product exists and has stock available
         if (product == null || product.getStock() <= 0) {
             throw new RuntimeException("Not enough stock for product: " + productId);
         }
         else
         {
-        	 shoppingCartService.addProductToCart(productId); // Add product to the cart
+            shoppingCartService.addProductToCart(productId); // Add product to the cart
         }
-        return shoppingCartService.getShoppingCart(); // Return the updated cart
+        
+        // Return the updated cart
+        return shoppingCartService.getShoppingCart();
     }
     
-    // Method to get the current shopping cart
+    /**
+     * Retrieves the current shopping cart for the user.
+     *
+     * @return The current shopping cart.
+     */
     public ShoppingCart getShoppingCart() {
         return shoppingCartService.getShoppingCart();
     }
 
-    // Method to get the total price of the cart
-    public double getTotalPrice() {
+    /**
+     * Calculates and returns the total price of all products in the shopping cart.
+     *
+     * @return The total price of the items in the cart.
+     */
+    public BigDecimal getTotalPrice() {
         return shoppingCartService.getTotalPrice();
     }
 }
